Fix all linker errors — Linux native port now links successfully
Resolve ~659 undefined reference errors through systematic fixes:

- Fix function signature type mismatches (signed/unsigned, pointer types)
  in audio2, spellvisuals, gametrek, math, SubstructA, substructB,
  dialogue, dialogmode, potionCraft, ContPakData
- Add extern "C" linkage to InitProc in seed.cpp
- Remove inline from constructors that need cross-TU visibility
  (GuiAnimatorBlip, WidgetMenuWorldMap, WidgetSaveFile)
- Add missing base class virtual method implementations for
  GuiAnimator, GuiAnimatorF, BaseWidget, IInventory,
  WidgetContPakData, WidgetTrainShop, WidgetSaveFile
- Fix Actor::remove_flagged_playerdata namespace scoping
- Fix borg6_func_a parameter type (Borg6Header* → Borg6Data*)
- Remove inline from GuiAnimator destructor declaration
- Create platform/linux/stubs.cpp with ROM data globals and
  function stubs using safe headers + forward declarations
- Add --allow-multiple-definition linker flag for N64 decompilation
  headers that intentionally define globals

Result: 13MB ELF x86-64 binary links and produces non-zero output.

+++ src_pseudo/potionCraft.cpp
@@ -2,7 +2,7 @@
 #include "widgets/ItemDetails.h"
 #include "globals.h"
 
-void makePotionMenu(u32 partyInd){
+void makePotionMenu(u8 partyInd){
     WHANDLE->AddWidget(new WidgetPotionCraft(partyInd));
 }
